refactor(api-gen): correct inconsistent filename
One file name was using camelCase
