+++ bazel/api-gen/rendering/regions/remove-eslint-comments.ts
